requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
pre-commit==2.9.3
django-import-export==2.5.0
